moved files in components and deleted duplicate profile.js file; add bootstrap classes to login and signup pages
